When running jitsi videobridge in amazon, we connect to our XMPP servers through an ELB. When deploys of the XMPP server happen, the nodes are terminated but the connection to the ELB still stays up, even though the nodes behind it have been removed. This causes the videobridge to think that it still has a connection to the XMPP server. Although the connection still appears valid, nothing is there to respond to pings or pubsub messages so the videobridge gets into a bad state. Forcing the underlying connection to reset based on failed ping requests helped to solve this for us.
The jitsi videobridge pom.xml will need to be updated to use the new version of Jicoco once this is merged into master. I'm not sure of the proper way to make that happen together so please advise here. Also did not see any documentation for jicoco and would be happy to start some but it felt like since this is used in the videobridge that the documentation should go there. Please advise as to where I should document this new configuration property.

The new configuration property can be set in the sip-comminicator.properties by adding the property configPropertiesBase.RECONNECT_ON_PING_FAILURES and setting it to true. To maintain the original behavior is treated as false by default. Setting this value will cause the connection to be reset after the number of pings set in the ping threshold fail.