feat(ComponentBase): use StripedExecutorService
Use StripedExecutorService to control the parallelism of the XMPP
packets processing. By implementing the 'getStripeForPacket' method
a component can enforce serial processing for packets that belong to
a "stripe". This allows for example to use conference ID from the
ColibriIQ as a stripe and make sure that packets related to specific
conference are processed in order and only one at a time. Same thing can
be used in Jicofo and other components if needed.